DB && Inheritance Optimizations

+++ controllers/admin/SyncController.php
@@ -1,6 +1,6 @@
 <?php
 
-include_once dirname(__FILE__).'/../abstract/PSEgoiController.php';
+include_once dirname(__FILE__).'/../SmartMarketingBaseController.php';
 
 /**
  * @package controllers/admin/SyncController
@@ -11,35 +11,63 @@ class SyncController extends SmartMarketingBaseController
 	/**
 	 * Constructor
 	 */
-	public function __construct() 
+	public function __construct()
 	{
 		parent::__construct();
 
 		$this->bootstrap = true;
-		$this->meta_title = $this->l('E-goi Subscribers').' - '.$this->module->displayName;
+		$this->cfg = 0;
+		$this->meta_title = $this->l('E-goi Sync Contacts').' - '.$this->module->displayName;
 		
 		if (!$this->module->active)
 			Tools::redirectAdmin($this->context->link->getAdminLink('AdminHome'));
 		
 		$this->retrieveRoles();
 		$this->mapFieldsEgoi();
-		$this->syncronizeEgoi($_POST);
+		$this->syncronizeEgoi();
 	}
+
+	/**
+	 * Inject Dependencies
+	 * 
+	 * @return mixed
+	 */
+	public function setMedia()
+	{
+		$this->addJS($this->_path. '/views/assets/js/sync.js');
+
+		return parent::setMedia();
+	}
+
+	/**
+	 * Toolbar settings
+	 * 
+	 * @return void
+	 */
+	public function initPageHeaderToolbar()
+    {
+        parent::initPageHeaderToolbar();
+        $this->page_header_toolbar_btn['save-and-stay'] = array(
+		    'short' => $this->l('Save Settings'),
+		    'href' => '#',
+		    'desc' => $this->l('Save Settings'),
+		    'js' => $this->l('$( \'#action_add\' ).click();')
+		);
+    }
 
 	/**
 	 * Initiate content
 	 * 
 	 * @return void
 	 */
-	public function initContent() 
+	public function initContent()
 	{
 		parent::initContent();
-		
-		if($this->has_api_key) {
-			
-			$msg = '';
-			if(!empty($_POST)){
-				$msg = $this->saveSync();
+
+		if ($this->isValid()) {
+
+			if(!empty($_POST)) {
+				$this->saveSync();
 			}
 
 			$api = new SmartApi();
@@ -49,15 +77,15 @@ class SyncController extends SmartMarketingBaseController
 
 			$rq = Db::getInstance(_PS_USE_SQL_SLAVE_)
 					->getRow('SELECT * FROM '._DB_PREFIX_.'egoi where client_id!="" order by egoi_id DESC');
-			
-			if(!empty($rq)){
+
+			if(!empty($rq)) {
 				$list_id = $rq['list_id'];
 				$sync = $rq['sync'];
 				$track = $rq['track'];
 				$role = $rq['role'];
 			}
 
-			if(isset($list_id) && ($list_id)){
+			if(isset($list_id) && ($list_id)) {
 				
 				$this->assign('subscribers', $api->getSubscribersFromListId($list_id));
 				$this->assign('list_id', $list_id);
@@ -75,12 +103,12 @@ class SyncController extends SmartMarketingBaseController
 					'birth_date' => 'Birth Date'
 				);
 
-				foreach($api->getExtraFields($list_id) as $key => $extra_field){
+				foreach($api->getExtraFields($list_id) as $key => $extra_field) {
 					$egoi_fields['extra_'.$key] = $extra_field['NAME'];
 				}
 				
 				$option = '';
-				foreach($egoi_fields as $key => $field){
+				foreach($egoi_fields as $key => $field) {
 					$option .= '<option value='.$key.'>'.$field.'</option>'.PHP_EOL;
 				}
 
@@ -91,10 +119,8 @@ class SyncController extends SmartMarketingBaseController
 			}
 
 			$this->assign('token', $_GET['token']);
-			$this->assign('egoi_success_message', $msg);
+			$this->assign('content', $this->fetch('sync.tpl'));
 		}
-
-		$this->assign('content', $this->fetch('subs.tpl'));
 	}
 
 	/**
@@ -102,23 +128,16 @@ class SyncController extends SmartMarketingBaseController
 	 * 
 	 * @return mixed
 	 */
-	public function saveSync() 
+	protected function saveSync() 
 	{
 		if(isset($_POST['action_add']) && ($_POST['action_add'])) {
 
 			$list = $_POST['list'];
 			$sync = $_POST['enable'];
 			$role = $_POST['role'];
-			$track = $_POST['track'];
-			$estado = 1;
-			
-			$update = '';
-			$insert = '';
+			$track = isset($_POST['track']) ? $_POST['track'] : 1;
 
-			if($track == ''){
-				$track = 1;
-			}
-
+			// compare client ID -> API with DB
 			$api = new SmartApi();
 			$client_data = $api->getClientData();
 			$client = $client_data['CLIENTE_ID'];
@@ -126,26 +145,22 @@ class SyncController extends SmartMarketingBaseController
 			$res = Db::getInstance(_PS_USE_SQL_SLAVE_)
 						->getRow('SELECT * FROM '._DB_PREFIX_.'egoi WHERE client_id='.(int)$client);
 
-			$sql_insert = array(
+			$values = array(
 				'list_id' => (int)$list, 
 				'client_id' => (int)$client,
 				'sync' => (int)$sync,
 				'track' => (int)$track,
 				'role' => pSQL($role),
-				'estado' => (int)$estado,
+				'estado' => 1,
                 'total' => 0
 			);
 
-			if($res['client_id']) {
-				$where = "client_id = ".(int)$client;
-				$update = Db::getInstance()->update('egoi', $sql_insert, $where);
+			if(isset($res['client_id']) && ($res['client_id'])) {
+				$this->assign('success_message', $this->displaySuccess($this->l('Settings updated')));
+				return Db::getInstance()->update('egoi', $values, "client_id = ".(int)$client);
 			}else{
-				$insert = Db::getInstance()->insert('egoi', $sql_insert);
-			}
-
-			if($insert || $update){
-				$msg = 1;
-				return $msg;
+				$this->assign('success_message', $this->displaySuccess($this->l('Settings saved')));
+				return Db::getInstance()->insert('egoi', $values);
 			}
 
 		}else{
@@ -158,7 +173,7 @@ class SyncController extends SmartMarketingBaseController
 	 * 
 	 * @return void
 	 */
-	public function retrieveRoles()
+	protected function retrieveRoles()
 	{
 		$this->assign('roles', Group::getGroups(Context::getContext()->language->id, true));
 	}
@@ -168,50 +183,53 @@ class SyncController extends SmartMarketingBaseController
 	 * 
 	 * @return void
 	 */
-	public function mapFieldsEgoi() 
+	protected function mapFieldsEgoi() 
 	{
-		$id = isset($_POST["id_egoi"]) ? (int)$_POST["id_egoi"] : '';
-		$token = isset($_POST["token_egoi_api"]) ? (int)$_POST["token_egoi_api"] : '';
-		$ps = isset($_POST["ps"]) ? pSQL($_POST["ps"]) : '';
-		$egoi = isset($_POST["egoi"]) ? pSQL($_POST["egoi"]) : '';
+		if (!empty($_POST)) {
+			$id = isset($_POST["id_egoi"]) ? (int)$_POST["id_egoi"] : '';
+			$token = isset($_POST["token_egoi_api"]) ? (int)$_POST["token_egoi_api"] : '';
+			$ps = isset($_POST["ps"]) ? pSQL($_POST["ps"]) : '';
+			$egoi = isset($_POST["egoi"]) ? pSQL($_POST["egoi"]) : '';
 
-		if(($token) && ($ps) && ($egoi)){
+			if(($token) && ($ps) && ($egoi)){
 
-			$ps_name = pSQL($_POST["ps_name"]);
-			$egoi_name = pSQL($_POST["egoi_name"]);
-			$status = 1;
+				$ps_name = pSQL($_POST["ps_name"]);
+				$egoi_name = pSQL($_POST["egoi_name"]);
+				$status = 1;
 
-			$sql_exists = "SELECT id FROM "._DB_PREFIX_."egoi_map_fields WHERE ps='".$ps."' OR egoi='".$egoi."'";
-			$exists = Db::getInstance(_PS_USE_SQL_SLAVE_)->executeS($sql_exists);
-			$field_exist = isset($exists[0]) ? $exists[0] : false;
+				$sql_exists = "SELECT id FROM "._DB_PREFIX_."egoi_map_fields WHERE ps='".$ps."' OR egoi='".$egoi."'";
+				$exists = Db::getInstance(_PS_USE_SQL_SLAVE_)->executeS($sql_exists);
+				$field_exist = isset($exists[0]) ? $exists[0] : false;
 
-			if (!$field_exist){
+				if (!$field_exist){
 
-				// insert new mapped field
-				Db::getInstance()->insert('egoi_map_fields', array(
-					'ps' => $ps,
-					'ps_name' => $ps_name,
-					'egoi' => $egoi,
-					'egoi_name' => $egoi_name,
-					'status' => (int)$status
-				));
+					// insert new mapped field
+					Db::getInstance()->insert('egoi_map_fields', array(
+						'ps' => $ps,
+						'ps_name' => $ps_name,
+						'egoi' => $egoi,
+						'egoi_name' => $egoi_name,
+						'status' => (int)$status
+					));
 
-	         	Db::getInstance(_PS_USE_SQL_SLAVE_)
-	         				->getRow("SELECT * FROM "._DB_PREFIX_."egoi_map_fields order by id DESC");
+		         	Db::getInstance(_PS_USE_SQL_SLAVE_)
+		         				->getRow("SELECT * FROM "._DB_PREFIX_."egoi_map_fields order by id DESC");
 
-         		$this->assign('post_id', Db::getInstance()->Insert_ID());
-         		$this->assign('ps_name', $ps_name);
-         		$this->assign('egoi_name', $egoi_name);
+	         		//TODO - test this
+	         		$this->assign('post_id', Db::getInstance()->Insert_ID());
+	         		$this->assign('ps_name', $ps_name);
+	         		$this->assign('egoi_name', $egoi_name);
 
-         		echo $this->context->smarty->display(_PS_MODULE_DIR_.$this->name.'/views/templates/admin/fields.tpl');
-	        }
+	         		echo $this->context->smarty->display($this->_path.'/views/templates/admin/fields.tpl');
+		        }
 
-			exit;
+				exit;
 
-		}else if($id){
+			}else if($id) {
 
-			Db::getInstance()->delete('egoi_map_fields', 'id = '.(int)$id);
-			exit;
+				Db::getInstance()->delete('egoi_map_fields', 'id = '.(int)$id);
+				exit;
+			}
 		}
 	}
 
@@ -221,7 +239,7 @@ class SyncController extends SmartMarketingBaseController
 	 * @param $listID
 	 * @return int        
 	 */
-	private function getEgoiSubscribers($listID)
+	protected function getEgoiSubscribers($listID)
 	{
 		$count = 0;
 		$api = new SmartApi();
@@ -239,23 +257,24 @@ class SyncController extends SmartMarketingBaseController
 	/**
 	 * Join Subscribers count from DB and API
 	 * 
-	 * @param  array  $post
 	 * @return void
 	 */
-	public function syncronizeEgoi($post = array())
+	protected function syncronizeEgoi()
 	{
-		if(!empty($post)) {
-			if(isset($post['action']) && ($post['action'] == 'synchronize')) {
+		if(!empty($_POST)) {
+			if(isset($_POST['action']) && ($_POST['action'] == 'synchronize')) {
 
-				$total = array();
-				$exec = Db::getInstance(_PS_USE_SQL_SLAVE_)
-							->getRow('SELECT COUNT(*) AS CT FROM '._DB_PREFIX_.'customer WHERE active="1"');
+				if (isset($_POST['list']) && ($_POST['list'])) {
+					$total = array();
+					$exec = Db::getInstance(_PS_USE_SQL_SLAVE_)
+								->getRow('SELECT COUNT(*) AS CT FROM '._DB_PREFIX_.'customer WHERE active="1"');
 
-				$total[] = $this->getEgoiSubscribers($post['list']);
-				$total[] = $exec['CT'];
+					$total[] = $this->getEgoiSubscribers($_POST['list']);
+					$total[] = $exec['CT'];
 
-				echo json_encode($total);
-				exit;
+					echo json_encode($total);
+					exit;
+				}
 			}
 		}
 	}
